backend/pyproject.toml
[project]
name = "support-ticket-analyst"
version = "0.1.0"
description = ""
authors = [
    {name = "Alan Bohannon",email = "alanbohannon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.123.9,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]
